Change the way of co-location noise feature instances number determining
There is new flag parameter called 'ncf_proportional'. If it's true, then the number of the instances of given co-location noise feature is proportional to the number of instances of given feature, which are participating in co-locations. If it's false, then the number of the instances of every co-location noise feature is similar: co-location noise feature id is chosen randomly with uniform distribution